Fix format string type mismatch in src/capability.c.
The type 'long' was correct on most machines but not according to the C
standard and Windows 64-bit.

+++ src/capability.c
@@ -247,8 +247,8 @@ capability_index_stats(void (*cb)(const char *line, void *privdata), void *privd
 			cb(buf, privdata);
 		}
 
-		rb_snprintf(buf, sizeof buf, "'%s': remaining bits - %ld", idx->name,
-			    (sizeof(unsigned int) * 8) - (idx->highest_bit - 1));
+		rb_snprintf(buf, sizeof buf, "'%s': remaining bits - %u", idx->name,
+			    (unsigned int)((sizeof(unsigned int) * 8) - (idx->highest_bit - 1)));
 		cb(buf, privdata);
 	}
